tests moved up into a subdirectory of chalice, replicating its structure
